Time: 62 ms (36.38%) | Memory: 14.4 MB (35.55%) - LeetSync

+++ 959-3sum-with-multiplicity/3sum-with-multiplicity.cpp
@@ -1,5 +1,5 @@
 #define ll long long
-#define mod (ll)(1e9+7)
+#define mod (int)(1e9+7)
 class Solution {
 public:
     int threeSumMulti(vector<int>& arr, int target) {
